[refactor] 컴포넌트 페이지별 폴더구조로 정리
