+++ FrontEnd/my-car/src/mycar/components/completePage/AnimationMovingCar.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import styled, { keyframes } from 'styled-components';
 import carImage from './image-91.png';
-import { Heading1Medium } from '../../style/typo';
+import { Heading1Medium } from '../../../style/typo';
 
 const ModalBgDiv = styled.div`
   position: absolute;
