+++ FrontEnd/my-car/src/mycar/routers/Complete.js
@@ -1,6 +1,6 @@
-import CompleteTopBanner from '../components/CompleteTopBanner';
-import CompletedFunction from '../components/CompletedFunction';
-import AnimationMovingCar from '../components/AnimationMovingCar';
+import CompleteTopBanner from '../components/completePage/CompleteTopBanner';
+import CompletedFunction from '../components/completePage/CompletedFunction';
+import AnimationMovingCar from '../components/completePage/AnimationMovingCar';
 
 function Complete() {
   return (
