--- FrontEnd/my-car/src/mycar/util/SelectedIndex.js
@@ -1,6 +1,0 @@
-export function SelectedIndex({ userOptionID, optionInfo }) {
-  const optionIndex = optionInfo.findIndex(
-    (option) => option.id === userOptionID,
-  );
-  return optionIndex;
-}
